+++ docker/gcp_pudl_etl.sh
@@ -64,14 +64,13 @@ function run_pudl_etl() {
         --gcs-cache-path gs://internal-zenodo-cache.catalyst.coop \
         --etl-settings "$PUDL_SETTINGS_YML" \
         --live-dbs test/validate \
-    && pg_ctlcluster 15 dagster stop \
     && touch "$PUDL_OUTPUT/success"
 }
 
 function save_outputs_to_gcs() {
     echo "Copying outputs to GCP bucket $PUDL_GCS_OUTPUT" && \
     gsutil -m cp -r "$PUDL_OUTPUT" "$PUDL_GCS_OUTPUT" && \
-    rm "$PUDL_OUTPUT/success"
+    rm -f "$PUDL_OUTPUT/success"
 }
 
 function upload_to_dist_path() {
@@ -125,7 +124,7 @@ function zenodo_data_release() {
 function notify_slack() {
     # Notify pudl-deployment slack channel of deployment status
     echo "Notifying Slack about deployment status"
-    message='# `${BUILD_ID}` status\n\n'
+    message="${BUILD_ID} status\n\n"
     if [[ "$1" == "success" ]]; then
         message+=":large_green_circle: :sunglasses: :unicorn_face: :rainbow: deployment succeeded!! :partygritty: :database_parrot: :blob-dance: :large_green_circle:\n\n"
     elif [[ "$1" == "failure" ]]; then
@@ -197,6 +196,9 @@ ZENODO_SUCCESS=0
 # 2>&1 redirects stderr to stdout.
 run_pudl_etl 2>&1 | tee "$LOGFILE"
 ETL_SUCCESS=${PIPESTATUS[0]}
+
+# This needs to happen regardless of the ETL outcome:
+pg_ctlcluster 15 dagster stop 2>&1 | tee -a "$LOGFILE"
 
 save_outputs_to_gcs 2>&1 | tee -a "$LOGFILE"
 SAVE_OUTPUTS_SUCCESS=${PIPESTATUS[0]}