+++ docs/release_notes.rst
@@ -63,6 +63,15 @@ Data Cleaning
 * Added a notebook :mod:`devtools/debug-column-mapping.ipynb` to make debugging manual
   column maps for new datasets simpler and faster.
 
+Metadata Cleaning
+^^^^^^^^^^^^^^^^^
+
+* Made a description field mandatory for all instances of ``Field`` and ``Resource``.
+  Updated the FIELD_METADATA and RESOURCE_METADATA so that had a description.
+  :issue:`3224`, :pr:`3283`.
+* Removed fields that are not used in any tables and removed the xfail from the
+  ``test_defined_fields_are_used`` test. :issue:`3224`, :pr:`3283`.
+
 ---------------------------------------------------------------------------------------
 v2023.12.01
 ---------------------------------------------------------------------------------------
@@ -263,7 +272,8 @@ Data Coverage
   The newly accessible tables include:
 
   * :ref:`core_ferc714__respondent_id` (linking FERC-714 respondents to EIA utilities)
-  * :ref:`core_ferc714__hourly_demand_pa` (hourly electricity demand by planning area)
+  * :ref:`out_ferc714__hourly_planning_area_demand` (hourly electricity demand by
+    planning area)
   * :ref:`out_ferc714__respondents_with_fips` (annual respondents with county FIPS IDs)
   * :ref:`out_ferc714__summarized_demand` (annual demand for FERC-714 respondents)
 
@@ -369,7 +379,7 @@ Analysis
   (:ref:`out_eia861__compiled_geometry_balancing_authorities` and
   :ref:`out_eia861__compiled_geometry_utilities`), and the estimated total hourly
   electricity demand for each US state in
-  :ref:`out_ferc714__hourly_predicted_state_demand`. See :issue:`1973`
+  :ref:`out_ferc714__hourly_estimated_state_demand`. See :issue:`1973`
   and :pr:`2550`.
 
 Deprecations